correct C++ compiler target types for 64-bit hosts

+++ bld/plusplus/h/target16.h
@@ -71,8 +71,8 @@ typedef short           target_short;
 typedef unsigned short  target_ushort;
 typedef short           target_int;
 typedef unsigned short  target_uint;
-typedef long            target_long;
-typedef unsigned long   target_ulong;
+typedef int             target_long;
+typedef unsigned int    target_ulong;
 
 typedef unsigned int    target_size_t;
 typedef unsigned short  target_offset_t;
